requirements.txt
requests
python-dotenv
pyupbit
pandas
matplotlib
git+https://github.com/twopirllc/pandas-ta
ccxt